dev-requirements.txt
-r requirements.txt
coverage==6.3.2
pytest==7.1.1
pytest-aiohttp==1.0.4
pytest-mock==3.7.0
pytest-cov==3.0.0